# Loosen component dict type constraints