--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -10,7 +10,7 @@ using Statistics
 using Test
 using TimeZones
 
-using AxisSets: Pattern, constraintmap, dimpaths, flatten, validate
+using AxisSets: Pattern, KeyedArray, constraintmap, dimpaths, flatten, validate
 using Impute: ThresholdError
 
 @testset "AxisSets.jl" begin