Also, restrict components to `KeyedArray` rather than `XArray`. Reasons for the change.

1. Support empty `KeyedDataset()` constructor which requires loosened key/value type constraints to be useful.
2. We want to support variable types and depths for the keys. If you construct a `KeyedDataset` with only keys `(:train, :input, :load)` and `(:predict, :input, :load)` then adding `(:train, :output)` would error.
3. The `XArray` type (`Union`) was rather verbose and made setting up the correct dispatch harder. Since we don't have a use-case where we want `NamedDimsArray(KeyedArray(...))` I don't think it's worth supporting.

Closes #39