@rofinn 2021-04-04
Based on a couple our current use-cases this shouldn't have that big of an impact on performance because we're operating on relative few large components, so the type stability of the component arrays will have a larger impact on performance than the key tuples.

Julia 1.6 on macos

# Construction

```julia
@btime KeyedDataset(
           flatten([
               :train => [
                   :temp => KeyedArray([1.0 1.1; missing 2.2; 3.0 3.3]; time=1:3, id=[:a, :b]),
                   :load => KeyedArray([7.0 7.7; 8.0 missing; 9.0 9.9]; time=1:3, loc=[:x, :y]),
               ],
               :predict => [
                   :temp => KeyedArray([1.0 missing; 2.0 2.2; 3.0 3.3]; time=1:3, id=[:a, :b]),
                   :load => KeyedArray([7.0 7.7; 8.1 missing; 9.0 9.9]; time=1:3, loc=[:x, :y]),
               ]
           ])...,
           constraints = Pattern[(:__, :time), (:__, :loc), (:train, :__, :id)]
       );
```
Before: 397.795 μs (787 allocations: 37.68 KiB)
After: 397.598 μs (864 allocations: 41.12 KiB)

```julia
@btime Impute.filter($constrained_ds; dims=:id)
```
Before: 112.706 μs (883 allocations: 57.52 KiB)
After: 72.201 μs (794 allocations: 42.45 KiB)